+++ apps/client/app/(afterLogin)/my-study/_components/index.ts
@@ -1,6 +1,7 @@
 export { default as AssignmentStatusBox } from "./AssignmentStatusBox";
 export { default as AttendanceStatusBox } from "./AttendanceStatusBox";
 export { default as DailyTaskCarousel } from "./DailyTaskCarousel";
+export { default as DailyTaskItem } from "./DailyTaskItem";
 export { default as DailyTasks } from "./DailyTasks";
 export { default as EmptyStudy } from "./EmptyStudy";
 export { default as Header } from "./Header";
